+++ flexiapi/resources/views/admin/account/carddav/create.blade.php
@@ -15,7 +15,7 @@
         <div class="select large">
             <select name="carddav_id">
                 @foreach ($carddavServers as $carddavServer)
-                    <option value="{{ $carddavServer->id }}">{{ $carddavServer->name }}</option>
+                    <option value="{{ $carddavServer->id }}">{{ $carddavServer->name }} - {{ $carddavServer->uri }}</option>
                 @endforeach
             </select>
             <label for="carddav_id">{{ __('CardDav Server') }}</label>
@@ -28,7 +28,7 @@
             @include('parts.errors', ['name' => 'username'])
         </div>
             <div>
-            <input placeholder="Username" name="domain" type="text" value="{{ old('domain') }}" required>
+            <input placeholder="domain.com, realm.tld" name="domain" type="text" value="{{ old('domain') }}" required>
             <label for="domain">{{ __('Domain') }}</label>
             @include('parts.errors', ['name' => 'domain'])
         </div>